Sync resid-fp with VICE's version, switching output from float to short int
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplay-libs-2.1.1@158 75748be7-0387-4326-94c7-f41ce0883d2c

+++ sidplay/c64/c64xsid.h
@@ -82,7 +82,7 @@ public:
             m_sid->voice (num, mute);
     }
 
-    float *buffer() {
+    short *buffer() {
         return m_sid->buffer();
     }
     int bufferpos() {
